Update to support bazel and buck tests (including plugins)
Also to show bazel building a plugin, i added the bazel template to verify-status
gerrit plugin.

+++ jenkins/gerrit-buck-build-plugin.sh
@@ -11,19 +11,22 @@ git checkout -f gerrit/{branch}
 rm -rf plugins/{name}
 git read-tree -u --prefix=plugins/{name} origin/{branch}
 
-TARGETS=$(echo "{targets}" | sed -e 's/{{name}}/{name}/g')
+if [ -f 'BUCK' ]
+then
+  TARGETS=$(echo "{targets}" | sed -e 's/{{name}}/{name}/g')
 
-SOURCE_LEVEL=$(grep "source_level" .buckconfig || echo "source_level=7")
-. set-java.sh $(echo $SOURCE_LEVEL | cut -d '=' -f 2 | tr -d '[[:space:]]')
+  SOURCE_LEVEL=$(grep "source_level" .buckconfig || echo "source_level=7")
+  . set-java.sh $(echo $SOURCE_LEVEL | cut -d '=' -f 2 | tr -d '[[:space:]]')
 
-buck build -v 3 $TARGETS
+  buck build -v 3 $TARGETS
 
-for JAR in $(buck targets --show_output $TARGETS | awk '{{print $2}}')
-do
-    PLUGIN_VERSION=$(git describe  --always origin/{branch})
-    echo -e "Implementation-Version: $PLUGIN_VERSION" > MANIFEST.MF
-    jar ufm $JAR MANIFEST.MF && rm MANIFEST.MF
-    DEST_JAR=buck-out/gen/plugins/{name}/$(basename $JAR)
-    [ "$JAR" -ef "$DEST_JAR" ] || mv $JAR $DEST_JAR
-    echo "$PLUGIN_VERSION" > buck-out/gen/plugins/{name}/$(basename $JAR-version)
-done
+  for JAR in $(buck targets --show_output $TARGETS | awk '{{print $2}}')
+  do
+      PLUGIN_VERSION=$(git describe  --always origin/{branch})
+      echo -e "Implementation-Version: $PLUGIN_VERSION" > MANIFEST.MF
+      jar ufm $JAR MANIFEST.MF && rm MANIFEST.MF
+      DEST_JAR=buck-out/gen/plugins/{name}/$(basename $JAR)
+      [ "$JAR" -ef "$DEST_JAR" ] || mv $JAR $DEST_JAR
+      echo "$PLUGIN_VERSION" > buck-out/gen/plugins/{name}/$(basename $JAR-version)
+  done
+fi
